requirements.txt
schedule
requests
beautifulsoup4
pywhatkit
PyYAML